docker: refactor build
Use the cuda devel image as the base so we have all necessary headers to
build llama.cpp.

Make sure we build with LLAMA_NATIVE=OFF so no native processor features
are enabled since we'll use GPU anyway and this way our image will be
portable.

Use the runtime image as the image we run since we don't need all the
development stuff.

Python dependencies are installed with pip since it will honor our build
flags.

NB: llama.cpp will happily log a warning and move on LLAMA_CUBLAS=ON is
set but there are no cuda headers present in the system.