feat: add ability to use OCI per customer
This is slightly different than the OpenAI / AzuerOpenAI runners.
Instead, OCI credentials are configured globally for a given Skynet
runner, just like the local one, and if configured, requests shall be
sent to OCI rather than to the local running llama.

The ability to run without a local llm directly targetting OCI is still
retained.