feat(ttt) refactor job processing
Extract the logic to select the right llm to an "LLM selector" entity.
This simplifies the processor to only worry about the processing part,
and the whole selection is encapsulated in LLMSelector.

If / when we need to make model selection more complex than just decide
which processor to use, it will be contained here.

While working on this I also noticed a logic error when initializing the
OCI LLM: the max tokens was taken from the first processed job.

Contrary to my original assesment, the transformers download is done
just once, so we don't need to reuse the ChatOCIGenAI instance and can
set the max_tokens property appropriately.